Renamed customControlRenders.ts to canvasControlRenders.ts
