+++ composables/useCanvasText.ts
@@ -7,7 +7,7 @@ import {
   createRotateControlRender,
   createTrashControlRender,
   createBringToFrontControlRender,
-} from '@/utils/customControlRenders'
+} from '~/utils/canvasControlRenders'
 import { getResizeImage, getRotateImage, getTrashCanImage, getBringToFrontImage } from '@/utils/customImageIcons'
 import { toggleObjectZOrder, setTextboxTextRadius } from '~/utils/canvasUtils'
 import { CircularTextbox } from '~/utils/circularTextbox'
